+++ integrations/aspnetcore/src/Scalar.AspNetCore/Converters/DocumentDownloadTypeJsonConverter.cs
@@ -1,0 +1,16 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace Scalar.AspNetCore;
+
+internal sealed class DocumentDownloadTypeJsonConverter : JsonConverter<DocumentDownloadType>
+{
+    public override DocumentDownloadType Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) =>
+        // We don't have to implement this method because we don't need to deserialize the DocumentDownloadType enum.
+        default;
+
+    public override void Write(Utf8JsonWriter writer, DocumentDownloadType value, JsonSerializerOptions options)
+    {
+        writer.WriteStringValue(value.ToStringFast(true));
+    }
+}