+++ integrations/aspnetcore/src/Scalar.AspNetCore/Options/Authentication/Flows/ClientCredentialsFlow.cs
@@ -21,6 +21,5 @@ public sealed class ClientCredentialsFlow : OAuthFlow
     /// Gets or sets the location where authentication credentials should be placed in HTTP requests.
     /// </summary>
     [JsonPropertyName("x-scalar-credentials-location")]
-    [JsonConverter(typeof(CredentialsLocationConverter))]
     public CredentialsLocation? CredentialsLocation { get; set; }
 }