+++ integrations/aspnetcore/src/Scalar.AspNetCore/ScalarExtensions/Badge.cs
@@ -1,0 +1,10 @@
+namespace Scalar.AspNetCore;
+
+internal sealed class Badge
+{
+    public required string Name { get; init; }
+
+    public required BadgePosition? Position { get; init; }
+
+    public required string? Color { get; init; }
+}