+++ packages/api-reference/src/components/Section/CompactSection.vue
@@ -7,21 +7,23 @@ import { Anchor } from '@/components/Anchor'
 import { Section } from '@/components/Section'
 import { useNavState } from '@/hooks/useNavState'
 
-const props = defineProps<{
+const { id, defaultOpen = false } = defineProps<{
   id: string
   label?: string
+  /** Control the initial open state of the section */
+  defaultOpen?: boolean
 }>()
 const { hash } = useNavState()
 
-const open = ref(false)
+const open = ref(defaultOpen)
 
 watch(
   hash,
-  async (id) => {
-    if (id === props.id && !open.value) {
+  async (newHash) => {
+    if (id === newHash && !open.value) {
       open.value = true
       await nextTick()
-      scrollToId(props.id)
+      scrollToId(id)
     }
   },
   { immediate: true },
@@ -38,9 +40,9 @@ watch(
       type="button"
       @click="open = !open">
       <ScalarIconCaretRight
-        weight="bold"
         class="size-3 transition-transform duration-100"
-        :class="{ 'rotate-90': open }" />
+        :class="{ 'rotate-90': open }"
+        weight="bold" />
       <Anchor
         :id="id"
         class="collapsible-section-header">