+++ integrations/docusaurus/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/docusaurus
 
+## 0.7.13
+
+### Patch Changes
+
+- Updated dependencies [50032be]
+  - @scalar/types@0.2.13
+
+## 0.7.12
+
+### Patch Changes
+
+- Updated dependencies [6a88108]
+  - @scalar/types@0.2.12
+
 ## 0.7.11
 
 ### Patch Changes