+++ integrations/fastapi/CHANGELOG.md
@@ -1,5 +1,15 @@
 # scalar-fastapi
 
+## 1.3.0
+
+### Minor Changes
+
+- 77af6e7: feat: use new HTML/JS API
+
+### Patch Changes
+
+- 6e22556: Add type marker according to PEP-561
+
 ## 1.2.3
 
 ### Patch Changes